requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
asn1crypto==0.22.0        # via cryptography
babel==2.4.0              # via flower
backoff==1.4.3
backports-abc==0.5        # via tornado
backports.ssl-match-hostname==3.5.0.1  # via tornado
bcrypt==3.1.3             # via paramiko
billiard==3.3.0.23        # via celery
boto3==1.4.4
boto==2.48.0
botocore==1.5.80          # via boto3, s3transfer
cachetools==2.0.0         # via google-auth
celery==3.1.25.pqueue     # via third_party dir
certifi==2017.4.17        # via requests, tornado
cffi==1.10.0              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
cryptography==1.9         # via paramiko
decorator==4.0.11         # via networkx
dill==0.2.7               # via google-gax
docutils==0.13.1          # via botocore
easywebdav==1.2.0
enum-compat==0.0.2        # via eventlet
enum34==1.1.6             # via cryptography, enum-compat, google-cloud-vision, grpcio
eventlet==0.21.0
fabric==1.13.2
filechunkio==1.8
flower==0.9.2
future==0.16.0            # via google-gax
futures==3.1.1            # via flower, grpcio, s3transfer
gapic-google-cloud-datastore-v1==0.15.3  # via google-cloud-datastore
gapic-google-cloud-error-reporting-v1beta1==0.15.3  # via google-cloud-error-reporting
gapic-google-cloud-logging-v2==0.91.3  # via google-cloud-logging
gapic-google-cloud-pubsub-v1==0.15.4  # via google-cloud-pubsub
gapic-google-cloud-spanner-admin-database-v1==0.15.3  # via google-cloud-spanner
gapic-google-cloud-spanner-admin-instance-v1==0.15.3  # via google-cloud-spanner
gapic-google-cloud-spanner-v1==0.15.3  # via google-cloud-spanner
gapic-google-cloud-speech-v1==0.15.3  # via google-cloud-speech
gevent==1.2.2
google-auth-httplib2==0.0.2  # via google-cloud-core
google-auth==1.0.1        # via google-auth-httplib2, google-cloud-core, google-cloud-storage, google-gax
google-cloud-bigquery==0.25.0  # via google-cloud
google-cloud-bigtable==0.25.0  # via google-cloud
google-cloud-core==0.25.0  # via google-cloud, google-cloud-bigquery, google-cloud-bigtable, google-cloud-datastore, google-cloud-dns, google-cloud-error-reporting, google-cloud-language, google-cloud-logging, google-cloud-monitoring, google-cloud-pubsub, google-cloud-resource-manager, google-cloud-runtimeconfig, google-cloud-spanner, google-cloud-speech, google-cloud-storage, google-cloud-translate, google-cloud-vision
google-cloud-datastore==1.1.0  # via google-cloud
google-cloud-dns==0.25.0  # via google-cloud
google-cloud-error-reporting==0.25.1  # via google-cloud
google-cloud-language==0.25.0  # via google-cloud
google-cloud-logging==1.1.0  # via google-cloud, google-cloud-error-reporting
google-cloud-monitoring==0.25.0
google-cloud-pubsub==0.26.0  # via google-cloud
google-cloud-resource-manager==0.25.0  # via google-cloud
google-cloud-runtimeconfig==0.25.0  # via google-cloud
google-cloud-spanner==0.25.0  # via google-cloud
google-cloud-speech==0.26.0  # via google-cloud
google-cloud-storage==1.2.0  # via google-cloud
google-cloud-translate==0.25.0  # via google-cloud
google-cloud-videointelligence==0.25.0  # via google-cloud
google-cloud-vision==0.25.0  # via google-cloud
google-cloud==0.26.1
google-gax==0.15.13       # via gapic-google-cloud-datastore-v1, gapic-google-cloud-error-reporting-v1beta1, gapic-google-cloud-logging-v2, gapic-google-cloud-pubsub-v1, gapic-google-cloud-spanner-admin-database-v1, gapic-google-cloud-spanner-admin-instance-v1, gapic-google-cloud-spanner-v1, gapic-google-cloud-speech-v1, google-cloud-bigtable, google-cloud-datastore, google-cloud-videointelligence, google-cloud-vision
google-resumable-media==0.1.1  # via google-cloud-storage
googleapis-common-protos[grpc]==1.5.2  # via gapic-google-cloud-datastore-v1, gapic-google-cloud-error-reporting-v1beta1, gapic-google-cloud-logging-v2, gapic-google-cloud-pubsub-v1, gapic-google-cloud-spanner-admin-database-v1, gapic-google-cloud-spanner-admin-instance-v1, gapic-google-cloud-spanner-v1, gapic-google-cloud-speech-v1, google-cloud-core, google-cloud-videointelligence, google-cloud-vision, google-gax, grpc-google-iam-v1, proto-google-cloud-datastore-v1, proto-google-cloud-error-reporting-v1beta1, proto-google-cloud-logging-v2, proto-google-cloud-pubsub-v1, proto-google-cloud-spanner-admin-database-v1, proto-google-cloud-spanner-admin-instance-v1, proto-google-cloud-spanner-v1, proto-google-cloud-speech-v1
greenlet==0.4.12          # via eventlet, gevent
grpc-google-iam-v1==0.11.1  # via gapic-google-cloud-pubsub-v1, gapic-google-cloud-spanner-admin-database-v1, gapic-google-cloud-spanner-admin-instance-v1
grpcio==1.4.0             # via google-cloud-logging, google-cloud-pubsub, google-cloud-spanner, google-cloud-speech, google-gax, googleapis-common-protos, grpc-google-iam-v1, proto-google-cloud-datastore-v1, proto-google-cloud-error-reporting-v1beta1, proto-google-cloud-logging-v2, proto-google-cloud-pubsub-v1, proto-google-cloud-spanner-admin-database-v1, proto-google-cloud-spanner-admin-instance-v1, proto-google-cloud-spanner-v1, proto-google-cloud-speech-v1
httplib2==0.10.3
idna==2.5                 # via cryptography, requests
ipaddress==1.0.18         # via cryptography
jmespath==0.9.3           # via boto3, botocore
kombu==3.0.37             # via celery
lxml==3.8.0
meld3==1.0.2              # via supervisor
msgpack-python==0.4.8
networkx==1.11            # via prov
oauth2client==3.0.0       # via gapic-google-cloud-datastore-v1, gapic-google-cloud-error-reporting-v1beta1, gapic-google-cloud-logging-v2, gapic-google-cloud-pubsub-v1, gapic-google-cloud-spanner-admin-database-v1, gapic-google-cloud-spanner-admin-instance-v1, gapic-google-cloud-spanner-v1, gapic-google-cloud-speech-v1, grpc-google-iam-v1, proto-google-cloud-datastore-v1, proto-google-cloud-error-reporting-v1beta1, proto-google-cloud-logging-v2, proto-google-cloud-pubsub-v1, proto-google-cloud-spanner-admin-database-v1, proto-google-cloud-spanner-admin-instance-v1, proto-google-cloud-spanner-v1, proto-google-cloud-speech-v1
paramiko==2.2.1           # via fabric
ply==3.8                  # via google-gax
proto-google-cloud-datastore-v1[grpc]==0.90.4  # via gapic-google-cloud-datastore-v1
proto-google-cloud-error-reporting-v1beta1[grpc]==0.15.3  # via gapic-google-cloud-error-reporting-v1beta1
proto-google-cloud-logging-v2[grpc]==0.91.3  # via gapic-google-cloud-logging-v2
proto-google-cloud-pubsub-v1[grpc]==0.15.4  # via gapic-google-cloud-pubsub-v1
proto-google-cloud-spanner-admin-database-v1[grpc]==0.15.3  # via gapic-google-cloud-spanner-admin-database-v1
proto-google-cloud-spanner-admin-instance-v1[grpc]==0.15.3  # via gapic-google-cloud-spanner-admin-instance-v1
proto-google-cloud-spanner-v1[grpc]==0.15.3  # via gapic-google-cloud-spanner-v1
proto-google-cloud-speech-v1[grpc]==0.15.3  # via gapic-google-cloud-speech-v1
protobuf==3.3.0
prov==1.3.1               # via prov-es
prov_es==0.1.1
psutil==5.2.2
pyasn1-modules==0.0.9     # via google-auth, google-auth-httplib2, oauth2client
pyasn1==0.2.3             # via google-auth, google-auth-httplib2, oauth2client, paramiko, pyasn1-modules, rsa
pycparser==2.18           # via cffi
pynacl==1.1.2             # via paramiko
python-dateutil==2.6.1    # via botocore, prov
pytz==2017.2              # via babel, celery, flower
redis==2.10.5
requests==2.18.1
rsa==3.4.2                # via google-auth, google-auth-httplib2, oauth2client
s3transfer==0.1.10        # via boto3
singledispatch==3.4.0.3   # via tornado
six==1.10.0               # via bcrypt, cryptography, google-auth, google-auth-httplib2, google-cloud-core, google-cloud-videointelligence, google-resumable-media, grpcio, oauth2client, protobuf, prov, pynacl, python-dateutil, singledispatch
supervisor==3.3.2
tornado==4.5.1            # via flower
urllib3==1.21.1           # via requests
